`CLOCConnector` and `CLOCProperties` refactoring
