+++ src/main/resources/application.properties
@@ -85,6 +85,9 @@ spring.gson.serialize-nulls=true
 ghs.git.folder-prefix=ghs-clone-
 ghs.git.clone-timeout-duration=5m
 
+# CLOC Configuration
+ghs.cloc.analysis-timeout-duration=5m
+
 # GitHub Configuration
 ghs.github.tokens=
 ghs.github.api-version=2022-11-28
@@ -119,7 +122,6 @@ ghs.crawler.delay-between-runs=PT6H
 ghs.analysis.enabled=true
 ghs.analysis.delay-between-runs=PT6H
 ghs.analysis.max-pool-threads=3
-ghs.analysis.cloc.analysis-timeout-duration=5m
 
 # Cleanup Configuration
 ghs.clean-up.enabled=true
